Extend the testing compiler to report errors
In order to test the generated errors we had to modify the existing
testing compiler and its interface. More precisely, we have to allow for a second
kind of error reporting that doesn't just bail out on errors but stores them
and allows for their retrieval.
Modified the Java interface accordingly, also needed to provide an additional
interface to store the error's position and message.